+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Things under "main" are not released yet.
 ## main
 
 - Allow prefixing a var with `_` to silence the unused variable warning.
+- Improve typechecking for string subscripts.
 
 ## v0.9.7